+++ src/pages/field/index.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import Test from "@Components/field/__test__";
+
+function TestPage() {
+  return (
+    <React.Fragment>
+      <Test />
+    </React.Fragment>
+  );
+}
+
+export default TestPage;